requirements.txt
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
PyMuPDF==1.25.3
pypdfium2==4.30.1
pytesseract==0.3.13
